Estilos .scss gral -> .scss componente. Refactorizacion validator. Mejoras de tipado (return:void)
