+++ src/app/features/pages/auth/register-form/register-form.component.ts
@@ -1,25 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
-
-//Fx valida contraseñas iguales
-export function MustMatch(controlName: string, matchingControlName: string) {
-  return (formGroup: FormGroup) => {
-      const control = formGroup.controls[controlName];
-      const matchingControl = formGroup.controls[matchingControlName];
-
-      if (matchingControl.errors && !matchingControl.errors.mustMatch) {
-          // return si otra validacion ya dio error
-          return;
-      }
-
-      // setea error en matchingControl si falla la validacion
-      if (control.value !== matchingControl.value) {
-          matchingControl.setErrors({ mustMatch: true });
-      } else {
-          matchingControl.setErrors(null);
-      }
-  }
-}
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { MustMatch } from 'src/app/shared/validators/pass-match.validator';
 
 @Component({
   selector: 'app-register-form',
@@ -29,15 +10,15 @@ export function MustMatch(controlName: string, matchingControlName: string) {
 
 export class RegisterFormComponent implements OnInit {
   
-  registerForm!: FormGroup
-  submitted = false;
+  registerForm!: FormGroup;
+  submitted: boolean = false;
 
   constructor(private formBuilder: FormBuilder) { }
 
   ngOnInit(): void { 
     this.registerForm = this.formBuilder.group({      
       email: ['', [Validators.required, Validators.email, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+[.][a-z]{2,4}')]],
-      password: ['', [Validators.required, Validators.minLength(6), Validators.pattern('(?=[^A-Z]*[A-Z])(?=[^a-z]*[a-z])(?=[^0-9]*[0-9])(?=.*[@$!%*#?&]).{6,}') ]],
+      password: ['', [Validators.required, Validators.minLength(6), Validators.pattern('(?=[^a-zA-Z]*[a-zA-Z])(?=[^0-9]*[0-9])(?=.*[@$!%*#?&]).{3,}') ]],
       confirmPassword: ['', Validators.required],
   }, {
       validator: MustMatch('password', 'confirmPassword')
@@ -47,7 +28,7 @@ export class RegisterFormComponent implements OnInit {
   // getter para rapido acceso a form fields
   get f() { return this.registerForm.controls; }
 
-  register(){
+  register(): void{
     console.log(this.registerForm);
     this.submitted = true;
 
@@ -57,7 +38,7 @@ export class RegisterFormComponent implements OnInit {
     }
   }
 
-  onReset() {
+  onReset(): void {
     this.submitted = false;
     this.registerForm.reset();
   }
