+++ src/app/shared/validators/pass-match.validator.ts
@@ -1,0 +1,20 @@
+import { FormGroup } from "@angular/forms";
+
+export function MustMatch(controlName: string, matchingControlName: string) {
+    return (formGroup: FormGroup) => {
+        const control = formGroup.controls[controlName];
+        const matchingControl = formGroup.controls[matchingControlName];
+  
+        if (matchingControl.errors && !matchingControl.errors.mustMatch) {
+            // return si otra validacion ya dio error
+            return;
+        }
+  
+        // setea error en matchingControl si falla la validacion
+        if (control.value !== matchingControl.value) {
+            matchingControl.setErrors({ mustMatch: true });
+        } else {
+            matchingControl.setErrors(null);
+        }
+    }
+  }
